add auth middleware
